Adding comments on functions.
