+++ smart-contracts/contracts/cl-vault/src/instantiate.rs
@@ -80,7 +80,7 @@ pub fn handle_instantiate(
         deps.storage,
         &deps.api.addr_validate(&msg.auto_compound_admin)?,
     )?;
-    MIGRATION_STATUS.save(deps.storage, &MigrationStatus::Open)?;
+    MIGRATION_STATUS.save(deps.storage, &MigrationStatus::Closed)?;
 
     let create_denom_msg: CosmosMsg = MsgCreateDenom {
         sender: env.contract.address.to_string(),
